+++ wrappers/dotnet/indy-shared-rs-dotnet/Models/RevocationRegistryDeltaValue.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 
 namespace indy_shared_rs_dotnet.Models
@@ -12,9 +13,9 @@ namespace indy_shared_rs_dotnet.Models
         public string Accum { get; set; }
 
         [JsonProperty("issued")]
-        public HashSet<uint> Issued { get; set; }
+        public HashSet<IntPtr> Issued { get; set; }
 
         [JsonProperty("revoked")]
-        public HashSet<uint> Revoked { get; set; }
+        public HashSet<IntPtr> Revoked { get; set; }
     }
 }
