+++ wrappers/dotnet/indy-shared-rs-dotnet/models/CredentialDefinition.cs
@@ -1,10 +1,11 @@
 ﻿using Newtonsoft.Json;
+using System;
 
 namespace indy_shared_rs_dotnet.Models
 {
     public class CredentialDefinition
     {
-        public uint Handle { get; set; }
+        public IntPtr Handle { get; set; }
 
         [JsonProperty("id")]
         public string CredentialDefinitionId { get; set; }
